Initial rewrite and cleanup
I no longer use a MQTT server, buzzer, and magnetic door switch so remove all related functions.
LED.show now runs on CORE 1 instead of CORE 0, this was causing the flickering in the LEDs all this time

Still need to cleanup a bit but it works with one LED effect and with my IOT app.